requirements.txt
discord.py[voice]
ffmpeg
aiohttp
PyNaCl
yt-dlp
google-api-python-client
python-dotenv